Move new classes to separate files
