+++ changelog.d/7481.feature
@@ -1,1 +1,1 @@
-Allow server admins to list users's devices and logout specific devices.
+Add admin APIs to allow server admins to list users' devices, and log out specific devices.
