+++ synapse/rest/admin/users.py
@@ -286,144 +286,6 @@ class UserRestServletV2(RestServlet):
             return 201, ret
 
 
-class DeviceRestServlet(RestServlet):
-    PATTERNS = (
-        re.compile(
-            "^/_synapse/admin/v2/users/(?P<user_id>[^/]*)/devices/(?P<device_id>[^/]*)$"
-        ),
-    )
-
-    """
-    Get, update or delete the given user's device
-    """
-
-    def __init__(self, hs):
-        """
-        Args:
-            hs (synapse.server.HomeServer): server
-        """
-        super(DeviceRestServlet, self).__init__()
-        self.hs = hs
-        self.auth = hs.get_auth()
-        self.device_handler = hs.get_device_handler()
-        self.store = hs.get_datastore()
-
-    async def on_GET(self, request, user_id, device_id):
-        await assert_requester_is_admin(self.auth, request)
-
-        target_user = UserID.from_string(user_id)
-        if not self.hs.is_mine(target_user):
-            raise SynapseError(400, "Can only lookup local users")
-
-        u = await self.store.get_user_by_id(target_user.to_string())
-        if u is None:
-            raise NotFoundError("Unknown user")
-
-        device = await self.device_handler.get_device(
-            target_user.to_string(), device_id
-        )
-        return 200, device
-
-    async def on_DELETE(self, request, user_id, device_id):
-        await assert_requester_is_admin(self.auth, request)
-
-        target_user = UserID.from_string(user_id)
-        if not self.hs.is_mine(target_user):
-            raise SynapseError(400, "Can only lookup local users")
-
-        u = await self.store.get_user_by_id(target_user.to_string())
-        if u is None:
-            raise NotFoundError("Unknown user")
-
-        await self.device_handler.delete_device(target_user.to_string(), device_id)
-        return 200, {}
-
-    async def on_PUT(self, request, user_id, device_id):
-        await assert_requester_is_admin(self.auth, request)
-
-        target_user = UserID.from_string(user_id)
-        if not self.hs.is_mine(target_user):
-            raise SynapseError(400, "Can only lookup local users")
-
-        u = await self.store.get_user_by_id(target_user.to_string())
-        if u is None:
-            raise NotFoundError("Unknown user")
-
-        body = parse_json_object_from_request(request, allow_empty_body=True)
-        await self.device_handler.update_device(
-            target_user.to_string(), device_id, body
-        )
-        return 200, {}
-
-
-class DevicesRestServlet(RestServlet):
-    PATTERNS = (re.compile("^/_synapse/admin/v2/users/(?P<user_id>[^/]*)/devices$"),)
-
-    """
-    Retrieve the given user's devices
-    """
-
-    def __init__(self, hs):
-        """
-        Args:
-            hs (synapse.server.HomeServer): server
-        """
-        self.hs = hs
-        self.auth = hs.get_auth()
-        self.device_handler = hs.get_device_handler()
-        self.store = hs.get_datastore()
-
-    async def on_GET(self, request, user_id):
-        await assert_requester_is_admin(self.auth, request)
-
-        target_user = UserID.from_string(user_id)
-        if not self.hs.is_mine(target_user):
-            raise SynapseError(400, "Can only lookup local users")
-
-        u = await self.store.get_user_by_id(target_user.to_string())
-        if u is None:
-            raise NotFoundError("Unknown user")
-
-        devices = await self.device_handler.get_devices_by_user(target_user.to_string())
-        return 200, {"devices": devices}
-
-
-class DeleteDevicesRestServlet(RestServlet):
-    """
-    API for bulk deletion of devices. Accepts a JSON object with a devices
-    key which lists the device_ids to delete.
-    """
-
-    PATTERNS = (
-        re.compile("^/_synapse/admin/v2/users/(?P<user_id>[^/]*)/delete_devices$"),
-    )
-
-    def __init__(self, hs):
-        self.hs = hs
-        self.auth = hs.get_auth()
-        self.device_handler = hs.get_device_handler()
-        self.store = hs.get_datastore()
-
-    async def on_POST(self, request, user_id):
-        await assert_requester_is_admin(self.auth, request)
-
-        target_user = UserID.from_string(user_id)
-        if not self.hs.is_mine(target_user):
-            raise SynapseError(400, "Can only lookup local users")
-
-        u = await self.store.get_user_by_id(target_user.to_string())
-        if u is None:
-            raise NotFoundError("Unknown user")
-
-        body = parse_json_object_from_request(request, allow_empty_body=False)
-        assert_params_in_dict(body, ["devices"])
-
-        await self.device_handler.delete_devices(
-            target_user.to_string(), body["devices"]
-        )
-        return 200, {}
-
-
 class UserRegisterServlet(RestServlet):
     """
     Attributes:
